requirements.txt
fastapi==0.124.4
uvicorn[standard]==0.38.0
python-dotenv==1.0.1
openai==2.14.0
cohere==5.20.0
qdrant-client==1.16.1
httpx==0.28.1
pydantic==2.12.4
python-multipart==0.0.18